requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.30.0
openai>=1.50.0
python-dotenv>=1.0.0
aiosqlite>=0.20.0
pytest>=8.0.0
pytest-asyncio>=0.24.0
